Configure CI tests, bumpversion
