+++ .bumpversion.cfg
@@ -1,0 +1,13 @@
+[bumpversion]
+current_version = 2.1.2
+commit = True
+tag = True
+tag_name = {new_version}
+
+[bumpversion:file:README.md]
+search = version = "{current_version}"
+replace = version = "{new_version}"
+
+[bumpversion:file:locals.tf]
+search = module_version = "{current_version}"
+replace = module_version = "{new_version}"
