+++ .github/workflows/terraform-CD.yml
@@ -18,7 +18,7 @@ env:
 jobs:
   publish:
     name: 'Publish Module'
-    runs-on: ubuntu-latest
+    runs-on: ["self-hosted", "Linux", "noble"]
     environment: production
     timeout-minutes: 60
     # Use the Bash shell regardless whether the GitHub Actions runner is ubuntu-latest, macos-latest, or windows-latest
