+++ locals.tf
@@ -1,5 +1,7 @@
 locals {
-  tags = {
+  module_version = "2.1.2"
+
+    tags = {
     created_by_module : "infrahouse/state-bucket/aws"
   }
 }
