+++ requirements-ci.txt
@@ -1,5 +1,2 @@
-boto3 ~= 1.26
-infrahouse-toolkit ~= 2.4
-pyhcl ~= 0.4
-yamllint ~= 1.29
-pytest ~= 7.4
+pytest-infrahouse ~= 0.9
+infrahouse-core ~= 0.17, >= 0.17.1
