+++ requirements.txt
@@ -1,3 +1,1 @@
-black ~= 23.1
-Sphinx ~= 6.1
 -r requirements-ci.txt
