+++ test_data/state-bucket/terraform.tfvars
@@ -1,0 +1,4 @@
+
+region = "us-west-1"
+
+role_arn = "arn:aws:iam::303467602807:role/state-bucket-tester"
