+++ test_data/state-bucket/variables.tf
@@ -1,0 +1,4 @@
+variable "region" {}
+variable "role_arn" {
+  default = null
+}
